+++ src/main/java/br/com/fiap/kessler_service/dto/CreditoOrbitalResponseDto.java
@@ -4,7 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -13,7 +12,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 public class CreditoOrbitalResponseDto {
     private Long idCredito;
-    private BigDecimal valorBrl;
+    private Double valorBrl;
     private String tipoComprador;
     private LocalDateTime dataTransacao;
     private String statusCredito;
